(feat/client): renomeando socket.py
